+++ contracts/hub/orderbook_deposits/src/lib.rs
@@ -3,7 +3,6 @@ mod error;
 pub mod execute;
 pub mod helpers;
 pub mod migrate;
-pub mod msg;
 pub mod query;
 pub mod state;
