+++ packages/euclid/src/msgs/mod.rs
@@ -5,6 +5,7 @@ pub mod factory;
 pub mod hook;
 pub mod lp_token;
 pub mod meta_transaction;
+pub mod orderbook_deposits;
 pub mod router;
 pub mod virtual_balance;
 pub mod vlp;
