+++ packages/euclid/src/msgs/orderbook_deposits/migrate.rs
@@ -1,0 +1,4 @@
+use cosmwasm_schema::cw_serde;
+
+#[cw_serde]
+pub struct MigrateMsg {}
