+++ packages/euclid/src/msgs/orderbook_deposits/mod.rs
@@ -1,0 +1,9 @@
+pub mod execute;
+pub mod instantiate;
+pub mod migrate;
+pub mod query;
+
+pub use execute::*;
+pub use instantiate::*;
+pub use migrate::*;
+pub use query::*;
